split GossipStats to separate file
